ao3_tags/annotate/requirements.txt
pandas==1.5.0
pyarrow==17.0.0
pyyaml==6.0.1
tqdm==4.66.1